fix: resolve OK-60815 and OK-60427 UI issues
